BSE: Resampler2: merge bseresamplerimpl.hh into bseresampler.cc
